AI/requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-chroma>=0.2.0
chromadb>=0.5.0
streamlit>=1.35.0
pydantic>=2.7.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.30.0
sse-starlette>=2.1.0
